Revert spectral fitting refactoring to address performance issues
REVERTED TO: commit b309a64 (just before initial refactoring)

Changes made:
- ✅ Restored original src/spectral_fitting.py (single file)
- ✅ Removed refactored src/spectral_fitting/ module
- ✅ Removed src/spectral_fitting_legacy.py
- ✅ Cleaned up debug logging additions
- ✅ Removed performance profiling tools
- ✅ Reverted rt_alignment.py timing additions
- ✅ Reverted run_jmod.py debug logger integration
- ✅ Removed debug logging command line options

PRESERVED:
- ✅ SPECTRAL_FITTING_REFACTORING_SUMMARY.md (comprehensive documentation)
- ✅ spectral_fitting_refactoring_plan.md (original plan)
- ✅ spectral_fitting_refactored_backup branch (full refactored code)

The refactoring work is fully documented and preserved in git history.
This revert should restore the original performance characteristics
while maintaining all the valuable refactoring insights for future use.

Tests: All imports work correctly after revert.
