+++ src/run_jmod.py
@@ -22,7 +22,7 @@ from .utils.misc_functions import write_to_csv
 from .utils import iso_functions as iso_f
 from .mass_tags import tag_library, available_tags
 from .fdr_analysis import process_data
-from .utils.debug_logger import setup_debug_logger, get_debug_logger, set_global_sample_rate, get_all_sampling_summaries
+# Removed debug logger imports - reverted to pre-refactoring state
 
 def main():
     """Main function to run JMod analysis."""
@@ -105,23 +105,7 @@ def main():
     if not os.path.exists(results_folder_path):
         os.mkdir(results_folder_path)
     
-    # Initialize debug logger if enabled
-    # if config.args.debug_log:
-    #     # Set the global sample rate
-    #     set_global_sample_rate(config.args.debug_sample_rate)
-    #     
-    #     logger = setup_debug_logger(results_folder_path, config.args.debug_log_level)
-    #     main_logger = get_debug_logger('run_jmod')
-    #     main_logger.info(f"Starting JMod analysis")
-    #     main_logger.info(f"Results folder: {results_folder_path}")
-    #     main_logger.info(f"Configuration: {config.args}")
-    #     main_logger.info(f"Debug sampling rate: {config.args.debug_sample_rate} ({config.args.debug_sample_rate * 100:.3f}% of messages)")
-    #     print(f"Debug logging enabled. Writing to: {results_folder_path}/debug.log")
-    #     print(f"Debug sampling rate: 1 in {int(1/config.args.debug_sample_rate)} messages")
-    # else:
-    #     # Disable debug logging
-    #     from .utils.debug_logger import set_debug_enabled
-    #     set_debug_enabled(False)
+    # Debug logging functionality removed - reverted to pre-refactoring state
     
     overall_start_time = time.time()
     #python run_jmod.py -r -l /Users/nathanwamsley/Data/SPEC_LIBS/JD_LF_Feb2025/LF_HY_lib.tsv -i /Users/nathanwamsley/Data/mzML/mTRAQ_Feb2025/JD0324.mzML --iso --num_iso 5
@@ -312,15 +296,7 @@ def main():
                  mass_tag=mass_tag,
                  timeplex=config.args.timeplex)
     
-    # Print debug sampling summary if debug logging was enabled
-    # if config.args.debug_log:
-    #     main_logger = get_debug_logger('run_jmod')
-    #     main_logger.info("Debug sampling summary:")
-    #     summaries = get_all_sampling_summaries()
-    #     if summaries:
-    #         main_logger.info(summaries)
-    #         print("\nDebug sampling summary:")
-    #         print(summaries)
+    # Debug sampling functionality removed - reverted to pre-refactoring state
     
     # """
