--- src/spectral_fitting/feature_calculation.py
@@ -1,577 +1,0 @@
-"""
-Feature calculation functions for spectral fitting.
-
-This module extracts and modularizes all feature calculations from the monolithic
-get_features() function, improving testability and maintainability.
-"""
-
-from typing import List, Tuple, Optional, Dict, Any
-import numpy as np
-from scipy import sparse
-import re
-
-from .types import (
-    SpectrumMatrix, BasicFeatures, SimilarityMetrics, 
-    StatisticalFeatures, FragmentInfo, SpectralFeatures
-)
-from ..utils.spectral_similarity_metrics import (
-    get_scribe, get_residuals, gof_stat, get_manhattan_distance
-)
-from ..utils.misc_functions import hyperscore_b_y, longest_y, cosim
-
-# Temporary timing imports for debugging
-from ..utils.timing_debug import time_function, Timer
-
-
-@time_function("feature_calc.basic_features")
-def calculate_basic_features(
-    spectrum_matrix: SpectrumMatrix,
-    dia_spectrum: np.ndarray,
-    coefficients: np.ndarray,
-    sparse_matrix: sparse.coo_matrix
-) -> BasicFeatures:
-    """
-    Calculate basic spectral matching features.
-    
-    Args:
-        spectrum_matrix: Unified spectrum matrix
-        dia_spectrum: DIA spectrum [mz, intensity]
-        coefficients: Fitted coefficients from NNLS
-        sparse_matrix: Sparse matrix used for fitting
-        
-    Returns:
-        BasicFeatures containing number of peaks matched, intensity fractions, etc.
-    """
-    # Get split data for each peptide
-    peptide_data_split = []
-    for idx in range(len(spectrum_matrix.peptide_candidates)):
-        row_idx, col_idx, values = spectrum_matrix.get_peptide_indices(idx)
-        peptide_data_split.append((row_idx, col_idx, values))
-    
-    # Number of peaks matched per peptide
-    num_peaks_matched = np.array([len(data[0]) for data in peptide_data_split])
-    
-    # Fraction of library intensity matched
-    frac_lib_intensity = np.array([np.sum(data[2]) for data in peptide_data_split])
-    
-    # Total intensity in DIA spectrum
-    tic = np.sum(dia_spectrum[:, 1])
-    
-    # Fraction of DIA intensity matched by each peptide
-    frac_dia_intensity = np.array([
-        np.sum(dia_spectrum[data[0][data[0] < len(dia_spectrum)], 1]) / tic if len(data[0]) > 0 and tic > 0 else 0
-        for data in peptide_data_split
-    ])
-    
-    # Get unique matched peaks
-    unique_row_idxs = np.unique(spectrum_matrix.row_indices)
-    if len(unique_row_idxs) > 0:
-        # Filter indices to be within bounds
-        valid_idxs = unique_row_idxs[unique_row_idxs < len(dia_spectrum)]
-        if len(valid_idxs) > 0:
-            dia_spec_int = dia_spectrum[valid_idxs, 1]
-            frac_int_matched_scalar = np.sum(dia_spec_int) / np.sum(dia_spectrum[:, 1])
-        else:
-            frac_int_matched_scalar = 0
-    else:
-        frac_int_matched_scalar = 0
-    
-    # Predicted spectrum from fitted coefficients
-    predicted_spec = np.squeeze(sparse_matrix * coefficients)[:-1]
-    frac_int_pred_scalar = np.sum(predicted_spec) / tic if tic > 0 else 0
-    
-    # Create arrays with same value for all peptides
-    num_peptides = len(spectrum_matrix.peptide_candidates)
-    frac_int_matched = np.full(num_peptides, frac_int_matched_scalar)
-    frac_int_pred = np.full(num_peptides, frac_int_pred_scalar)
-    
-    return BasicFeatures(
-        num_peaks_matched=num_peaks_matched,
-        frac_lib_intensity=frac_lib_intensity,
-        frac_dia_intensity=frac_dia_intensity,
-        frac_int_matched=frac_int_matched,
-        frac_int_pred=frac_int_pred
-    )
-
-
-def calculate_similarity_metrics(
-    spectrum_matrix: SpectrumMatrix,
-    dia_spectrum: np.ndarray,
-    coefficients: np.ndarray,
-    sparse_matrix: sparse.coo_matrix,
-    ref_spec_offset: int,
-    decoy_spec_offset: int
-) -> SimilarityMetrics:
-    """
-    Calculate spectral similarity metrics.
-    
-    Args:
-        spectrum_matrix: Unified spectrum matrix
-        dia_spectrum: DIA spectrum [mz, intensity]
-        coefficients: Fitted coefficients
-        sparse_matrix: Sparse matrix used for fitting
-        ref_spec_offset: Offset for reference spectra in coefficients
-        decoy_spec_offset: Offset for decoy spectra in coefficients
-        
-    Returns:
-        SimilarityMetrics containing SCRIBE scores, Manhattan distances, etc.
-    """
-    # Get split data for calculations
-    row_indices_split = []
-    col_indices_split = []
-    values_split = []
-    
-    for idx in range(len(spectrum_matrix.peptide_candidates)):
-        row_idx, col_idx, values = spectrum_matrix.get_peptide_indices(idx)
-        # Filter row indices to be within bounds of dia_spectrum
-        valid_mask = row_idx < len(dia_spectrum)
-        row_indices_split.append(row_idx[valid_mask])
-        col_indices_split.append(col_idx[valid_mask])
-        values_split.append(values[valid_mask])
-    
-    # SCRIBE scores
-    scribe_scores = get_scribe(
-        row_indices_split,
-        col_indices_split,
-        values_split,
-        dia_spectrum[:, 1]
-    )
-    
-    # Get residuals and predictions
-    target_matrix = spectrum_matrix.get_target_data()
-    decoy_matrix = spectrum_matrix.get_decoy_data() if np.any(spectrum_matrix.is_decoy) else None
-    
-    # Split data for residual calculation
-    target_row_split = []
-    target_col_split = []
-    target_val_split = []
-    decoy_row_split = []
-    decoy_col_split = []
-    decoy_val_split = []
-    
-    for idx, is_decoy in enumerate(spectrum_matrix.is_decoy):
-        row_idx, col_idx, values = spectrum_matrix.get_peptide_indices(idx)
-        # Filter out indices that are out of bounds
-        valid_mask = row_idx < len(dia_spectrum)
-        row_idx = row_idx[valid_mask]
-        col_idx = col_idx[valid_mask]
-        values = values[valid_mask]
-        
-        if is_decoy and decoy_matrix is not None:
-            decoy_row_split.append(row_idx)
-            decoy_col_split.append(col_idx)
-            decoy_val_split.append(values)
-        else:
-            target_row_split.append(row_idx)
-            target_col_split.append(col_idx)
-            target_val_split.append(values)
-    
-    residuals, y_pred = get_residuals(
-        target_val_split,
-        target_row_split,
-        target_col_split,
-        decoy_val_split,
-        decoy_row_split,
-        decoy_col_split,
-        dia_spectrum[:, 1],
-        coefficients,
-        ref_spec_offset,
-        decoy_spec_offset
-    )
-    
-    # Manhattan distances
-    manhattan_distances, fitted_spectral_contrasts = get_manhattan_distance(
-        row_indices_split,
-        col_indices_split,
-        values_split,
-        dia_spectrum[:, 1],
-        y_pred
-    )
-    
-    # Correlation metrics (simplified for now)
-    num_peptides = len(spectrum_matrix.peptide_candidates)
-    r2_all = np.zeros(num_peptides)
-    r2_lib_spec = np.zeros(num_peptides)
-    r2_unique = np.zeros(num_peptides)
-    
-    # Cosine similarity for large coefficients
-    large_coeff_indices = np.where(coefficients > 1)[0]
-    if len(large_coeff_indices) > 0:
-        # Scale matrix by coefficients
-        large_coeffs = np.squeeze(coefficients)
-        large_coeffs[large_coeffs < 1] = 0
-        scaled_matrix = np.multiply(sparse_matrix.toarray(), large_coeffs)
-        subset_pred_spec = np.sum(scaled_matrix, 1)
-        
-        # Get subset of matched rows
-        subset_row_indices = np.unique(
-            sparse_matrix.row[np.isin(sparse_matrix.col, large_coeff_indices)]
-        )
-        subset_row_indices = subset_row_indices[subset_row_indices < len(dia_spectrum)]
-        
-        if len(subset_row_indices) > 0:
-            dia_spec_int = dia_spectrum[subset_row_indices, 1]
-            subset_cosine = cosim(dia_spec_int, subset_pred_spec[subset_row_indices])
-        else:
-            subset_cosine = 0
-    else:
-        subset_cosine = 0
-    
-    cosine_similarity = np.full(num_peptides, subset_cosine)
-    
-    return SimilarityMetrics(
-        scribe_scores=scribe_scores,
-        manhattan_distances=manhattan_distances,
-        fitted_spectral_contrasts=fitted_spectral_contrasts,
-        r2_all=r2_all,
-        r2_lib_spec=r2_lib_spec,
-        r2_unique=r2_unique,
-        cosine_similarity=cosine_similarity
-    )
-
-
-def calculate_statistical_features(
-    spectrum_matrix: SpectrumMatrix,
-    dia_spectrum: np.ndarray,
-    coefficients: np.ndarray,
-    residuals: np.ndarray,
-    ref_spec_offset: int
-) -> StatisticalFeatures:
-    """
-    Calculate statistical analysis features.
-    
-    Args:
-        spectrum_matrix: Unified spectrum matrix
-        dia_spectrum: DIA spectrum
-        coefficients: Fitted coefficients
-        residuals: Residuals from fitting
-        ref_spec_offset: Offset for reference spectra
-        
-    Returns:
-        StatisticalFeatures containing goodness of fit, residual analysis, etc.
-    """
-    # Get split data
-    row_indices_split = []
-    col_indices_split = []
-    values_split = []
-    
-    for idx in range(len(spectrum_matrix.peptide_candidates)):
-        row_idx, col_idx, values = spectrum_matrix.get_peptide_indices(idx)
-        # Filter row indices to be within bounds
-        valid_mask = row_idx < len(dia_spectrum)
-        row_indices_split.append(row_idx[valid_mask])
-        col_indices_split.append(col_idx[valid_mask]) 
-        values_split.append(values[valid_mask])
-    
-    # Goodness of fit statistics
-    gof_stats, max_unmatched_residuals, max_matched_residuals = gof_stat(
-        row_indices_split,
-        col_indices_split,
-        values_split,
-        residuals,
-        dia_spectrum[:, 1],
-        coefficients,
-        ref_spec_offset
-    )
-    
-    # Fraction predictions
-    frac_lib_intensity = np.array([np.sum(vals) for vals in values_split])
-    frac_dia_intensity = np.array([
-        np.sum(dia_spectrum[row_idx, 1]) / np.sum(dia_spectrum[:, 1]) 
-        if len(row_idx) > 0 else 0
-        for row_idx in row_indices_split
-    ])
-    
-    # Unique peak predictions
-    frac_unique_pred = np.zeros(len(spectrum_matrix.peptide_candidates))
-    for idx in range(len(spectrum_matrix.peptide_candidates)):
-        if idx >= ref_spec_offset and idx < ref_spec_offset + np.sum(~spectrum_matrix.is_decoy):
-            coeff = coefficients[idx]
-            if len(row_indices_split[idx]) > 0 and coeff > 0:
-                # Find peaks unique to this peptide
-                # (simplified - full implementation would check single matches)
-                frac_unique_pred[idx] = frac_lib_intensity[idx] * coeff
-    
-    # Fraction of DIA intensity predicted
-    frac_dia_intensity_pred = np.zeros_like(frac_dia_intensity)
-    for idx in range(len(spectrum_matrix.peptide_candidates)):
-        if idx >= ref_spec_offset and frac_dia_intensity[idx] > 0:
-            coeff_idx = idx - ref_spec_offset
-            if coeff_idx < len(coefficients):
-                frac_dia_intensity_pred[idx] = (
-                    frac_lib_intensity[idx] * coefficients[coeff_idx] / frac_dia_intensity[idx]
-                )
-    
-    return StatisticalFeatures(
-        gof_stats=gof_stats,
-        max_unmatched_residuals=max_unmatched_residuals,
-        max_matched_residuals=max_matched_residuals,
-        frac_unique_pred=frac_unique_pred,
-        frac_dia_intensity_pred=frac_dia_intensity_pred
-    )
-
-
-def calculate_fragment_features(
-    spectrum_matrix: SpectrumMatrix,
-    dia_spectrum: np.ndarray,
-    lib_peaks_matched: List[np.ndarray],
-    prec_frags: Optional[List[Dict[str, Any]]] = None,
-    ordered_frags: Optional[List[List[str]]] = None,
-    pep_cand_list: Optional[List[np.ndarray]] = None
-) -> FragmentInfo:
-    """
-    Calculate fragment-level features.
-    
-    Args:
-        spectrum_matrix: Unified spectrum matrix
-        dia_spectrum: DIA spectrum [mz, intensity]
-        lib_peaks_matched: Boolean arrays of matched library peaks
-        prec_frags: Fragment information dictionaries
-        ordered_frags: Ordered fragment names
-        pep_cand_list: Original peptide candidate spectra (optional)
-        
-    Returns:
-        FragmentInfo containing hyperscores, b/y counts, etc.
-    """
-    num_peptides = len(spectrum_matrix.peptide_candidates)
-    
-    if prec_frags and len(prec_frags) > 0:
-        if ordered_frags is not None:
-            # Use ordered fragments with new hyperscore calculation
-            hyperscores, b_counts, y_counts = map(list, zip(*[
-                hyperscore2(frags, frag_names) 
-                for frags, frag_names in zip(prec_frags, ordered_frags)
-            ]))
-            longest_y_ions = [
-                max([int(re.match(r"[by](\d+)", i)[1]) for i in frag_names])
-                for frag_names in ordered_frags
-            ]
-        else:
-            # Use original hyperscore calculation
-            hyperscores, b_counts, y_counts = map(list, zip(*[
-                hyperscore_b_y(frags, j) 
-                for frags, j in zip(prec_frags, lib_peaks_matched)
-            ]))
-            longest_y_ions = [
-                longest_y(frags, j) 
-                for frags, j in zip(prec_frags, lib_peaks_matched)
-            ]
-    else:
-        # No fragment information available
-        hyperscores = np.zeros(num_peptides)
-        b_counts = np.zeros(num_peptides)
-        y_counts = np.zeros(num_peptides)
-        longest_y_ions = np.zeros(num_peptides)
-    
-    # Fragment errors and m/z values
-    frag_errors = []
-    frag_mz = []
-    frag_names = []
-    frag_int = []
-    obs_int = []
-    
-    # Extract fragment information for each peptide
-    if lib_peaks_matched is not None and len(lib_peaks_matched) > 0:
-        # Get split data for each peptide
-        peptide_data_split = []
-        for idx in range(len(spectrum_matrix.peptide_candidates)):
-            row_idx, col_idx, values = spectrum_matrix.get_peptide_indices(idx)
-            peptide_data_split.append((row_idx, col_idx, values))
-        
-        # Process fragment information for each peptide
-        for idx in range(min(len(lib_peaks_matched), len(peptide_data_split))):
-            if idx < len(lib_peaks_matched):
-                matched = lib_peaks_matched[idx]
-                row_indices = peptide_data_split[idx][0]
-                values = peptide_data_split[idx][2]
-                
-                # Get matched fragments
-                if ordered_frags and idx < len(ordered_frags):
-                    matched_frag_names = [ordered_frags[idx][i] for i in range(len(ordered_frags[idx])) if i < len(matched) and matched[i]]
-                    frag_names.append(matched_frag_names)
-                else:
-                    frag_names.append([])
-                
-                # Fragment errors would need bin_centers from DIA spectrum processing
-                # For now, append empty lists
-                frag_errors.append([])
-                frag_mz.append([])
-                
-                # Get fragment intensities from prec_frags dictionary if available
-                if prec_frags and idx < len(prec_frags) and ordered_frags and idx < len(ordered_frags):
-                    # Extract intensities for matched fragments
-                    frag_dict = prec_frags[idx]
-                    frag_order = ordered_frags[idx]
-                    if frag_dict and len(frag_order) > 0 and len(matched_frag_names) > 0:
-                        # Get intensities for matched fragments
-                        matched_intensities = []
-                        for frag_name in matched_frag_names:
-                            if frag_name in frag_dict:
-                                matched_intensities.append(frag_dict[frag_name])
-                            else:
-                                matched_intensities.append(0.0)
-                        frag_int.append(matched_intensities)
-                    else:
-                        frag_int.append([])
-                else:
-                    # Fallback to using values from spectrum matrix
-                    if hasattr(matched, '__len__') and len(matched) <= len(values):
-                        frag_int.append(values[:len(matched)][matched])
-                    else:
-                        frag_int.append([])
-                
-                # Observed intensities from DIA spectrum
-                valid_row_indices = row_indices[row_indices < len(dia_spectrum)]
-                if len(valid_row_indices) > 0:
-                    obs_int.append(dia_spectrum[valid_row_indices, 1])
-                else:
-                    obs_int.append([])
-            else:
-                frag_names.append([])
-                frag_errors.append([])
-                frag_mz.append([])
-                frag_int.append([])
-                obs_int.append([])
-    
-    return FragmentInfo(
-        hyperscores=np.array(hyperscores),
-        b_counts=np.array(b_counts),
-        y_counts=np.array(y_counts),
-        longest_y_ions=np.array(longest_y_ions),
-        frag_errors=frag_errors,
-        frag_mz=frag_mz,
-        frag_names=frag_names,
-        frag_int=frag_int,
-        obs_int=obs_int
-    )
-
-
-def hyperscore2(frags: Dict[str, Any], frag_names_matched: List[str]) -> Tuple[float, int, int]:
-    """
-    Calculate hyperscore for matched fragments.
-    
-    Args:
-        frags: Fragment information
-        frag_names_matched: Names of matched fragments
-        
-    Returns:
-        Tuple of (hyperscore, b_count, y_count)
-    """
-    num_b = sum(["b" in i for i in frag_names_matched if "iso" not in i])
-    num_y = sum(["y" in i for i in frag_names_matched if "iso" not in i])
-    dp = np.sum([frags[i] for i in frag_names_matched if "iso" not in i])
-    
-    hyperscore = max(0, np.log(dp * np.math.factorial(num_b) * np.math.factorial(num_y)))
-    
-    return hyperscore, num_b, num_y
-
-
-@time_function("feature_calc.all_features")
-def calculate_all_features(
-    spectrum_matrix: SpectrumMatrix,
-    dia_spectrum: np.ndarray,
-    coefficients: np.ndarray,
-    sparse_matrix: sparse.coo_matrix,
-    rt_mz: np.ndarray,
-    window_idxs: np.ndarray,
-    prec_rt: float,
-    ms1_errors: np.ndarray,
-    ref_spec_offset: int = 0,
-    decoy_spec_offset: int = 0,
-    lib_peaks_matched: Optional[List[np.ndarray]] = None,
-    prec_frags: Optional[List[Dict[str, Any]]] = None,
-    ordered_frags: Optional[List[List[str]]] = None,
-    pep_cand_list: Optional[List[np.ndarray]] = None
-) -> List[SpectralFeatures]:
-    """
-    Calculate all spectral features for each peptide.
-    
-    This is the main entry point that combines all feature calculations.
-    
-    Args:
-        spectrum_matrix: Unified spectrum matrix
-        dia_spectrum: DIA spectrum
-        coefficients: Fitted coefficients
-        sparse_matrix: Sparse matrix used for fitting
-        rt_mz: RT and m/z values for candidates
-        window_idxs: Indices of candidates in window
-        prec_rt: Precursor retention time
-        ms1_errors: MS1 mass errors
-        ref_spec_offset: Offset for reference spectra
-        decoy_spec_offset: Offset for decoy spectra
-        lib_peaks_matched: Matched library peaks
-        prec_frags: Fragment information
-        ordered_frags: Ordered fragment names
-        pep_cand_list: Original peptide candidate spectra
-        
-    Returns:
-        List of SpectralFeatures for each peptide
-    """
-    # Calculate feature groups
-    basic = calculate_basic_features(spectrum_matrix, dia_spectrum, coefficients, sparse_matrix)
-    
-    similarity = calculate_similarity_metrics(
-        spectrum_matrix, dia_spectrum, coefficients, sparse_matrix,
-        ref_spec_offset, decoy_spec_offset
-    )
-    
-    # Get residuals for statistical features
-    residuals = np.zeros_like(dia_spectrum[:, 1])  # Placeholder
-    
-    statistical = calculate_statistical_features(
-        spectrum_matrix, dia_spectrum, coefficients, residuals, ref_spec_offset
-    )
-    
-    fragment_info = calculate_fragment_features(
-        spectrum_matrix, dia_spectrum, lib_peaks_matched or [], prec_frags, ordered_frags, pep_cand_list
-    )
-    
-    # RT and m/z errors
-    rt_errors = prec_rt - rt_mz[:, 0][window_idxs]
-    precursor_mzs = rt_mz[:, 1][window_idxs]
-    
-    # Create SpectralFeatures for each peptide
-    features_list = []
-    for idx in range(len(spectrum_matrix.peptide_candidates)):
-        features = SpectralFeatures(
-            basic=BasicFeatures(
-                num_peaks_matched=basic.num_peaks_matched[idx:idx+1],
-                frac_lib_intensity=basic.frac_lib_intensity[idx:idx+1],
-                frac_dia_intensity=basic.frac_dia_intensity[idx:idx+1],
-                frac_int_matched=basic.frac_int_matched[idx:idx+1],
-                frac_int_pred=basic.frac_int_pred[idx:idx+1]
-            ),
-            similarity=SimilarityMetrics(
-                scribe_scores=similarity.scribe_scores[idx:idx+1],
-                manhattan_distances=similarity.manhattan_distances[idx:idx+1],
-                fitted_spectral_contrasts=similarity.fitted_spectral_contrasts[idx:idx+1],
-                r2_all=similarity.r2_all[idx:idx+1],
-                r2_lib_spec=similarity.r2_lib_spec[idx:idx+1],
-                r2_unique=similarity.r2_unique[idx:idx+1],
-                cosine_similarity=similarity.cosine_similarity[idx:idx+1]
-            ),
-            statistical=StatisticalFeatures(
-                gof_stats=statistical.gof_stats[idx:idx+1],
-                max_unmatched_residuals=statistical.max_unmatched_residuals[idx:idx+1],
-                max_matched_residuals=statistical.max_matched_residuals[idx:idx+1],
-                frac_unique_pred=statistical.frac_unique_pred[idx:idx+1],
-                frac_dia_intensity_pred=statistical.frac_dia_intensity_pred[idx:idx+1]
-            ),
-            fragment_info=FragmentInfo(
-                hyperscores=fragment_info.hyperscores[idx:idx+1],
-                b_counts=fragment_info.b_counts[idx:idx+1],
-                y_counts=fragment_info.y_counts[idx:idx+1],
-                longest_y_ions=fragment_info.longest_y_ions[idx:idx+1],
-                frag_errors=fragment_info.frag_errors[idx:idx+1] if idx < len(fragment_info.frag_errors) else [],
-                frag_mz=fragment_info.frag_mz[idx:idx+1] if idx < len(fragment_info.frag_mz) else [],
-                frag_names=fragment_info.frag_names[idx:idx+1] if idx < len(fragment_info.frag_names) else [],
-                frag_int=fragment_info.frag_int[idx:idx+1] if idx < len(fragment_info.frag_int) else [],
-                obs_int=fragment_info.obs_int[idx:idx+1] if idx < len(fragment_info.obs_int) else []
-            ),
-            ms1_error=ms1_errors[idx:idx+1] if idx < len(ms1_errors) else np.array([0]),
-            rt_error=rt_errors[idx:idx+1] if idx < len(rt_errors) else np.array([0]),
-            precursor_mz=precursor_mzs[idx:idx+1] if idx < len(precursor_mzs) else np.array([0])
-        )
-        features_list.append(features)
-    
-    return features_list
