--- src/spectral_fitting/fitting_core.py
@@ -1,631 +1,0 @@
-"""
-Core fitting algorithms for spectral fitting.
-
-This module contains the unified fit_spectrum_to_library function that replaces
-the duplicated fit_to_lib and fit_to_lib_decoy functions.
-"""
-
-from typing import Dict, List, Tuple, Optional, Any, Union
-import numpy as np
-import warnings
-from scipy import sparse
-import logging
-
-from .types import (
-    SpectralFitResult, FittingParameters, PrecursorInfo,
-    SpectrumMatrix, SpectralFeatures
-)
-from .matrix_operations import (
-    create_spectrum_matrix, add_unmatched_peaks_to_matrix,
-    rank_and_create_sparse_matrix
-)
-from .feature_calculation import calculate_all_features
-from ..utils.misc_functions import (
-    createTolWindows, window_width, feature_list_mz, 
-    feature_list_rt, closest_ms1spec, closest_peak_diff
-)
-from ..utils.spectral_similarity_metrics import get_closest_ms1
-from ..utils.io.read_output import names
-from ..spectral_fitting_legacy import create_entries
-import src.config as config
-import ptinnls as sparse_nnls
-
-# Get logger for this module
-logger = logging.getLogger('jmod_debug.spectral_fitting.fitting_core')
-
-# Import sampling logger for high-frequency logs
-from ..utils.debug_logger import get_sampling_logger, get_global_sample_rate
-
-# Function to get the sampling logger with current rate
-def get_current_sampling_logger():
-    """Get a sampling logger with the current global sample rate."""
-    return get_sampling_logger('spectral_fitting.fitting_core', sample_rate=get_global_sample_rate())
-
-# This will be initialized when first used
-sampling_logger = None
-
-# Temporary timing imports for debugging
-from ..utils.timing_debug import time_function, Timer
-
-
-def prepare_dia_spectrum(
-    dia_spectrum: np.ndarray,
-    mz_tol: float
-) -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
-    """
-    Process DIA spectrum by merging peaks within tolerance.
-    
-    Args:
-        dia_spectrum: Raw DIA spectrum [mz, intensity]
-        mz_tol: m/z tolerance for merging peaks
-        
-    Returns:
-        Tuple of (processed_spectrum, centroid_breaks, bin_centers)
-    """
-    # Find first indices of peaks grouped by tolerance
-    merged_coords_idxs = np.searchsorted(
-        dia_spectrum[:, 0] + mz_tol * dia_spectrum[:, 0],
-        dia_spectrum[:, 0]
-    )
-    
-    # Get first mz of each peak group
-    merged_coords = dia_spectrum[np.unique(merged_coords_idxs), 0]
-    
-    # Sum intensities for each group
-    merged_intensities = np.zeros(len(merged_coords_idxs))
-    for j, val in zip(merged_coords_idxs, dia_spectrum[:, 1]):
-        merged_intensities[j] += val
-    merged_intensities = merged_intensities[merged_intensities != 0]
-    
-    # Update spectrum to merged values
-    processed_spectrum = np.array((merged_coords, merged_intensities)).transpose()
-    
-    # Calculate tolerance windows
-    centroid_breaks = np.concatenate((
-        processed_spectrum[:, 0] - mz_tol * processed_spectrum[:, 0],
-        processed_spectrum[:, 0] + mz_tol * processed_spectrum[:, 0]
-    ))
-    centroid_breaks = np.sort(centroid_breaks)
-    bin_centers = np.mean(np.stack((centroid_breaks[::2], centroid_breaks[1::2]), 1), 1)
-    
-    return processed_spectrum, centroid_breaks, bin_centers
-
-
-@time_function("fitting_core.filter_candidates")
-def filter_candidates_by_window(
-    rt_mz: np.ndarray,
-    prec_info: PrecursorInfo,
-    params: FittingParameters,
-    ms1_mz: Optional[float] = None,
-    dino_features: Optional[Any] = None
-) -> np.ndarray:
-    """
-    Filter peptide candidates based on m/z and RT windows.
-    
-    Args:
-        rt_mz: Array of [RT, m/z] for all candidates
-        prec_info: Precursor information
-        params: Fitting parameters
-        ms1_mz: Optional MS1 m/z for filtering
-        dino_features: Optional DINO features for filtering
-        
-    Returns:
-        Indices of candidates within window
-    """
-    # DEBUG: Window filtering - use sampling logger for high-frequency calls
-    # global sampling_logger
-    # if sampling_logger is None:
-    #     sampling_logger = get_current_sampling_logger()
-    
-    # sampling_logger.debug(f"filter_candidates: prec_mz={prec_info.mz:.4f}, rt={prec_info.rt:.2f}, window={prec_info.window_width:.4f}")
-    # sampling_logger.debug(f"  Total library entries: {len(rt_mz)}")
-    
-    if ms1_mz:
-        _bool = np.abs(rt_mz[:, 1] - ms1_mz) < params.ms1_tol
-        # sampling_logger.debug(f"  Using MS1 m/z filtering: {ms1_mz:.4f}, tol={params.ms1_tol}")
-    else:
-        if params.use_rt:
-            _bool = np.logical_and(
-                np.abs(rt_mz[:, 1] - prec_info.mz) < (prec_info.window_width / 2),
-                np.abs(rt_mz[:, 0] - prec_info.rt) < params.rt_tol
-            )
-            # sampling_logger.debug(f"  Using RT filtering: rt_tol={params.rt_tol}")
-        else:
-            _bool = np.abs(rt_mz[:, 1] - prec_info.mz) < (prec_info.window_width / 2)
-            # sampling_logger.debug(f"  RT filtering disabled")
-    
-    window_idxs = np.where(_bool)[0]
-    # sampling_logger.debug(f"  Candidates in m/z window: {len(window_idxs)}")
-    
-    # Additional filtering with DINO features if provided
-    if dino_features is not None:
-        filtered_dino = feature_list_mz(
-            feature_list_rt(dino_features, prec_info.rt, rt_tol=params.rt_tol),
-            prec_info.mz, prec_info.window_width
-        )
-        window_edges = createTolWindows(filtered_dino.mz, tolerance=params.ms1_tol)
-        before_dino = len(window_idxs)
-        window_idxs = window_idxs[
-            np.where((np.searchsorted(window_edges, rt_mz[window_idxs, 1]) % 2) == 1)[0]
-        ]
-        # sampling_logger.debug(f"  After DINO filtering: {before_dino} -> {len(window_idxs)}")
-    
-    # sampling_logger.debug(f"  Final window_idxs: {len(window_idxs)}")
-    return window_idxs
-
-
-@time_function("fitting_core.create_entries")
-def create_entries(
-    centroid_breaks: np.ndarray,
-    candidate_peaks: List[np.ndarray],
-    mass_window_candidates: List[Any],
-    top_n: int,
-    atleast_m: int,
-    prec_mzs: np.ndarray,
-    ms1_spec: Any,
-    ms1_tol: float,
-    frac_matched: float = 0.5,
-    spec_frags: Optional[List[Any]] = None,
-    top_n_idxs: Optional[List[np.ndarray]] = None
-) -> Tuple[List[int], List[Any], List[np.ndarray], List[np.ndarray], 
-           List[np.ndarray], List[np.ndarray], List[np.ndarray], 
-           List[np.ndarray], List[np.ndarray], np.ndarray]:
-    """
-    Create entries for peptides that meet matching criteria.
-    
-    This function filters peptides based on peak matching requirements and
-    prepares the data structures needed for fitting.
-    """
-    # DEBUG: Entry creation - use sampling logger
-    # sampling_logger.debug(f"create_entries: {len(candidate_peaks)} candidates")
-    # if len(candidate_peaks) > 0:
-    #     sampling_logger.debug(f"  First candidate has {len(candidate_peaks[0])} peaks")
-    # sampling_logger.debug(f"  Centroid breaks: {len(centroid_breaks)} breaks")
-    
-    # Find coordinates of library peaks in DIA spectrum
-    ref_coords = [np.searchsorted(centroid_breaks, M[:, 0]) for M in candidate_peaks]
-    
-    # Get top N peaks by intensity
-    if top_n_idxs is not None:
-        # Use pre-calculated indices with coords (like legacy version)
-        top_ten = [c[idxs] for c, idxs in zip(ref_coords, top_n_idxs)]
-    else:
-        top_ten = [
-            np.searchsorted(centroid_breaks, M[np.argsort(-M[:, 1])[0:min(top_n, M.shape[0])], 0])
-            for M in candidate_peaks
-        ]
-    
-    # Filter by MS1 peak presence
-    ms1_error = np.array([closest_peak_diff(mz, ms1_spec.mz, max_diff=ms1_tol) for mz in prec_mzs])
-    ms1_peak = ~np.isnan(ms1_error)
-    # sampling_logger.debug(f"  MS1 peaks found: {np.sum(ms1_peak)} / {len(ms1_peak)}")
-    
-    # Normalize intensities
-    all_norm_intensities = [M[:, 1] / sum(M[:, 1]) for M in candidate_peaks]
-    
-    # DEBUG: Show filtering criteria - use sampling logger
-    # sampling_logger.debug(f"  Filtering criteria: frac_lib_matched={config.frac_lib_matched}, atleast_m={atleast_m}")
-    # sampling_logger.debug(f"  Peaks passing intensity filter: {sum(1 for i in range(len(candidate_peaks)) if (all_norm_intensities[i][(ref_coords[i] % 2) == 1]).sum() > config.frac_lib_matched)}")
-    # sampling_logger.debug(f"  Peaks passing top_n filter: {sum(1 for i in range(len(candidate_peaks)) if (top_ten[i] % 2).sum() > atleast_m)}")
-    
-    # Filter peptides meeting criteria
-    if config.match_ms1:
-        ref_peaks_in_dia = [
-            i for i in range(len(candidate_peaks))
-            if ((all_norm_intensities[i][(ref_coords[i] % 2) == 1]).sum() > config.frac_lib_matched and
-                (top_ten[i] % 2).sum() > atleast_m and
-                ms1_peak[i])
-        ]
-    else:
-        ref_peaks_in_dia = [
-            i for i in range(len(candidate_peaks))
-            if ((all_norm_intensities[i][(ref_coords[i] % 2) == 1]).sum() > config.frac_lib_matched and
-                (top_ten[i] % 2).sum() > atleast_m)
-        ]
-    
-    # sampling_logger.debug(f"  Final peaks_in_dia: {len(ref_peaks_in_dia)}")
-    
-    # Show example of peak matching for first candidate - use sampling logger
-    # if len(candidate_peaks) > 0:
-    #     sampling_logger.debug(f"  Example peak matching for first candidate:")
-    #     sampling_logger.debug(f"    Library peaks (first 5): {candidate_peaks[0][:5, 0] if len(candidate_peaks[0]) > 0 else 'None'}")
-    #     sampling_logger.debug(f"    Coords (first 5): {ref_coords[0][:5] if len(ref_coords[0]) > 0 else 'None'}")
-    #     sampling_logger.debug(f"    Matched (first 5): {[c % 2 == 1 for c in ref_coords[0][:5]] if len(ref_coords[0]) > 0 else 'None'}")
-    #     sampling_logger.debug(f"    Total matched: {sum(1 for c in ref_coords[0] if c % 2 == 1)} / {len(ref_coords[0])}")
-    
-    # Extract data for filtered peptides
-    ref_pep_cand_loc = [ref_coords[i] for i in ref_peaks_in_dia]
-    ref_pep_cand_list = [candidate_peaks[i] for i in ref_peaks_in_dia]
-    ref_pep_cand = [mass_window_candidates[i] for i in ref_peaks_in_dia]
-    
-    norm_intensities = [M[:, 1] / sum(M[:, 1]) for M in ref_pep_cand_list]
-    
-    # Library peaks that match DIA peaks
-    lib_peaks_matched = [j % 2 == 1 for j in ref_pep_cand_loc]
-    
-    # Create split arrays for matrix construction
-    # Calculate row indices with bounds checking
-    # searchsorted can return values up to len(centroid_breaks)
-    # When converted to peak indices, this could exceed the valid range
-    ref_spec_row_indices_split = [
-        np.array([np.int32(min(((idx+1)/2)-1, (len(centroid_breaks)//2)-1)) for idx in i[j]], dtype=np.int32)
-        for i, j in zip(ref_pep_cand_loc, lib_peaks_matched)
-    ]
-    num_lib_peaks_matched = np.array([np.sum(i) for i in lib_peaks_matched])
-    ref_spec_col_indices_split = [
-        np.array([idx] * i, dtype=int) 
-        for idx, i in zip(range(len(ref_pep_cand)), num_lib_peaks_matched)
-    ]
-    ref_spec_values_split = [
-        ints[i] for ints, i in zip(norm_intensities, lib_peaks_matched)
-    ]
-    
-    # Get MS1 errors for filtered peaks
-    ref_ms1_error = ms1_error[ref_peaks_in_dia]
-    
-    return (ref_peaks_in_dia, ref_pep_cand, ref_pep_cand_loc, ref_pep_cand_list,
-            ref_spec_row_indices_split, ref_spec_col_indices_split,
-            ref_spec_values_split, norm_intensities, lib_peaks_matched, ref_ms1_error)
-
-
-@time_function("fitting_core.fit_spectrum_to_library")
-def fit_spectrum_to_library(
-    dia_spec: Any,
-    library: Dict[Union[int, str], Dict[str, Any]],
-    rt_mz: np.ndarray,
-    all_keys: List[Union[int, str]],
-    dino_features: Optional[Any] = None,
-    rt_filter: bool = False,
-    ms1_mz: Optional[float] = None,
-    ms1_spectra: Optional[List[Any]] = None,
-    mz_func: callable = np.array,
-    rt_tol: float = None,
-    ms1_tol: float = None,
-    mz_tol: float = None,
-    decoy_library: Optional[Dict[Union[int, str], Dict[str, Any]]] = None,
-    include_decoys: bool = True
-) -> SpectralFitResult:
-    """
-    Unified function for fitting a DIA spectrum to a spectral library.
-    
-    This function replaces both fit_to_lib and fit_to_lib_decoy, eliminating
-    900+ lines of code duplication. It handles both target and decoy peptides
-    in a single pass using the unified SpectrumMatrix data structure.
-    
-    Args:
-        dia_spec: DIA spectrum to fit
-        library: Spectral library
-        rt_mz: RT-m/z matrix for candidates
-        all_keys: Library keys
-        dino_features: Optional DINO features for filtering
-        rt_filter: Whether to apply RT filtering
-        ms1_mz: Optional MS1 m/z for filtering
-        ms1_spectra: Optional MS1 spectra
-        mz_func: Calibration function (default: identity)
-        rt_tol: RT tolerance (uses config default if None)
-        ms1_tol: MS1 tolerance (uses config default if None)
-        mz_tol: m/z tolerance (uses config default if None)
-        decoy_library: Optional separate decoy library
-        include_decoys: Whether to include decoy peptides
-        
-    Returns:
-        SpectralFitResult containing features for all peptides
-    """
-    # Use config defaults if tolerances not provided
-    if rt_tol is None:
-        rt_tol = config.rt_tol
-    if ms1_tol is None:
-        ms1_tol = config.ms1_tol
-    if mz_tol is None:
-        mz_tol = config.mz_tol
-    
-    # Get parameters from config
-    params = FittingParameters(
-        rt_tol=rt_tol,
-        mz_tol=mz_tol,
-        ms1_tol=ms1_tol,
-        top_n=config.top_n,
-        atleast_m=config.atleast_m,
-        frac_matched=0.5,
-        use_rt=rt_filter,
-        use_decoy=include_decoys
-    )
-    
-    # Extract spectrum information
-    spec = dia_spec
-    dia_spectrum = np.stack(spec.peak_list(), 1)
-    prec_info = PrecursorInfo(
-        mz=spec.prec_mz,
-        rt=spec.RT,
-        scan_num=spec.scan_num,
-        window_width=window_width(dia_spec)
-    )
-    
-    # Get MS1 spectrum if available
-    ms1_spec = None
-    if ms1_spectra is not None:
-        ms1_spec = get_closest_ms1(prec_info.rt, ms1_spectra)
-    
-    # Filter candidates by window
-    window_idxs = filter_candidates_by_window(
-        rt_mz, prec_info, params, ms1_mz, dino_features
-    )
-    
-    if len(window_idxs) == 0:
-        # No candidates found
-        return SpectralFitResult(
-            features=[],
-            coefficients=np.array([]),
-            peptide_ids=[],
-            is_decoy=np.array([], dtype=bool),
-            sparse_matrix=sparse.coo_matrix((0, 0)),
-            matched_peak_indices=np.array([])
-        )
-    
-    # Get candidate peptides and their spectra
-    with Timer("fitting_core.get_candidates"):
-        mass_window_candidates = [all_keys[i] for i in window_idxs]
-        candidate_peaks = [library[i]['spectrum'] for i in mass_window_candidates]
-    
-    # Process DIA spectrum
-    processed_spectrum, centroid_breaks, bin_centers = prepare_dia_spectrum(
-        dia_spectrum, params.mz_tol
-    )
-    
-    
-    # Get top N indices if available
-    top_n_idxs = [library[i].get('top_n') for i in mass_window_candidates]
-
-    spec_frags = None
-
-    if "spec_frags" in library[all_keys[0]]:
-        spec_frags = [library[i]['spec_frags'] for i in mass_window_candidates]
-    
-    # Create entries for target peptides
-    (ref_peaks_in_dia, ref_pep_cand, ref_pep_cand_loc, ref_pep_cand_list,
-     ref_spec_row_indices_split, ref_spec_col_indices_split,
-     ref_spec_values_split, norm_intensities, lib_peaks_matched, 
-     ref_ms1_error) = create_entries(
-        centroid_breaks, candidate_peaks, mass_window_candidates,
-        params.top_n, params.atleast_m, rt_mz[:, 1][window_idxs],
-        ms1_spec, params.ms1_tol, params.frac_matched,
-        spec_frags, top_n_idxs
-    )
-    
-    # Process decoy peptides if requested
-    decoy_spec_row_indices_split = []
-    decoy_spec_col_indices_split = []
-    decoy_spec_values_split = []
-    decoy_pep_cand = []
-    decoy_pep_cand_list = []
-    decoy_ms1_error = np.array([])
-    decoy_lib_peaks_matched = []
-    
-    if include_decoys and decoy_library is not None:
-        # Create decoy candidates
-        mass_window_decoy_candidates = [
-            ("Decoy_" + str(i[0]) if isinstance(i, tuple) else "Decoy_" + str(i), *i[1:]) 
-            if isinstance(i, tuple) else ("Decoy_" + str(i),)
-            for i in mass_window_candidates
-        ]
-        
-        # Get decoy data from library
-        converted_frags = [decoy_library[i]["frags"] for i in mass_window_candidates]
-        decoy_sorted_frags = [decoy_library[i]["ordered_frags"] for i in mass_window_candidates]
-        candidate_decoy_peaks = [decoy_library[i]["spectrum"] for i in mass_window_candidates]
-        decoy_mz = rt_mz[:, 1][window_idxs] - config.decoy_mz_offset
-        
-        decoy_top_n_idxs = [decoy_library[i].get('top_n') for i in mass_window_candidates]
-        
-        # Create entries for decoy peptides
-        (decoy_peaks_in_dia, decoy_pep_cand, decoy_pep_cand_loc, decoy_pep_cand_list,
-         decoy_spec_row_indices_split, decoy_spec_col_indices_split,
-         decoy_spec_values_split, norm_decoy_intensities, decoy_lib_peaks_matched,
-         decoy_ms1_error) = create_entries(
-            centroid_breaks, candidate_decoy_peaks, mass_window_decoy_candidates,
-            params.top_n, params.atleast_m, decoy_mz, ms1_spec, params.ms1_tol,
-            params.frac_matched, None, decoy_top_n_idxs
-        )
-    
-    # Create unified spectrum matrix
-    spectrum_matrix = create_spectrum_matrix(
-        ref_spec_row_indices_split,
-        ref_spec_col_indices_split,
-        ref_spec_values_split,
-        ref_pep_cand,
-        decoy_spec_row_indices_split,
-        decoy_spec_col_indices_split,
-        decoy_spec_values_split,
-        decoy_pep_cand
-    )
-    
-    # DEBUG: Matrix creation - use sampling logger
-    # sampling_logger.debug(f"  Matrix created: {len(spectrum_matrix.row_indices)} non-zero entries")
-    # sampling_logger.debug(f"  Unique rows: {len(np.unique(spectrum_matrix.row_indices)) if len(spectrum_matrix.row_indices) > 0 else 0}")
-    # sampling_logger.debug(f"  Total peptides: {len(spectrum_matrix.peptide_candidates)}")
-    
-    # Handle no matches case
-    if len(spectrum_matrix.row_indices) == 0 or len(ref_pep_cand) == 0:
-        return SpectralFitResult(
-            features=[],
-            coefficients=np.array([]),
-            peptide_ids=spectrum_matrix.peptide_candidates if len(spectrum_matrix.peptide_candidates) > 0 else [],
-            is_decoy=spectrum_matrix.is_decoy if len(spectrum_matrix.is_decoy) > 0 else np.array([], dtype=bool),
-            sparse_matrix=sparse.coo_matrix((1, 1)),  # Avoid empty matrix
-            matched_peak_indices=np.array([])
-        )
-    
-    # Get unique matched peak indices
-    unique_row_idxs = np.unique(spectrum_matrix.row_indices)
-    # Filter to valid indices
-    unique_row_idxs = unique_row_idxs[unique_row_idxs < len(processed_spectrum)]
-    dia_spec_int = processed_spectrum[unique_row_idxs, 1]
-    
-    # Add unmatched peaks
-    spectrum_matrix, _, _, _ = add_unmatched_peaks_to_matrix(
-        spectrum_matrix, norm_intensities, ref_pep_cand_loc,
-        fit_type=config.unmatched_fit_type
-    )
-    
-    # Create sparse matrix and fit
-    sparse_matrix, dia_spec_int = rank_and_create_sparse_matrix(
-        spectrum_matrix, dia_spec_int
-    )
-    
-    # Ensure dia_spec_int is the right length for the sparse matrix
-    if sparse_matrix.shape[0] > len(dia_spec_int):
-        dia_spec_int = np.append(dia_spec_int, [0] * (sparse_matrix.shape[0] - len(dia_spec_int)))
-    elif sparse_matrix.shape[0] < len(dia_spec_int):
-        dia_spec_int = dia_spec_int[:sparse_matrix.shape[0]]
-    
-    # Convert to dense array if needed for ptinnls
-    if hasattr(sparse_matrix, 'toarray'):
-        sparse_matrix_dense = sparse_matrix.toarray()
-    else:
-        sparse_matrix_dense = sparse_matrix
-    
-    # Ensure dia_spec_int is a column vector
-    dia_spec_int = dia_spec_int.reshape(-1)
-    
-    # Perform sparse NNLS fitting
-    with Timer("fitting_core.nnls_solve"):
-        with warnings.catch_warnings():
-            warnings.simplefilter("ignore")
-            # Check if we have a valid matrix
-            if sparse_matrix_dense.shape[0] == 0 or sparse_matrix_dense.shape[1] == 0:
-                lib_coefficients = np.array([])
-            else:
-                fit_results = sparse_nnls.lsqnonneg(sparse_matrix_dense, dia_spec_int, {"show_progress": False})
-                lib_coefficients = fit_results['x']
-                # Convert to numpy array if it's a cvxopt matrix
-                if hasattr(lib_coefficients, '__array__'):
-                    lib_coefficients = np.array(lib_coefficients).flatten()
-                else:
-                    lib_coefficients = np.asarray(lib_coefficients).flatten()
-    
-    # Calculate all features
-    ref_spec_offset = 0
-    decoy_spec_offset = np.sum(~spectrum_matrix.is_decoy)
-    
-    # If no coefficients, return empty result
-    if len(lib_coefficients) == 0:
-        return SpectralFitResult(
-            features=[],
-            coefficients=lib_coefficients,
-            peptide_ids=spectrum_matrix.peptide_candidates,
-            is_decoy=spectrum_matrix.is_decoy,
-            sparse_matrix=sparse_matrix,
-            matched_peak_indices=unique_row_idxs
-        )
-    
-    # Extract fragment information from library for feature calculation
-    prec_frags = []
-    ordered_frags = []
-    for pep_id in ref_pep_cand:
-        if pep_id in library:
-            if "frags" in library[pep_id]:
-                prec_frags.append(library[pep_id]["frags"])
-            else:
-                prec_frags.append({})
-                
-            if "ordered_frags" in library[pep_id]:
-                ordered_frags.append(library[pep_id]["ordered_frags"])
-            else:
-                ordered_frags.append([])
-        else:
-            prec_frags.append({})
-            ordered_frags.append([])
-    
-    # Combine with decoy fragments if present
-    all_lib_peaks_matched = lib_peaks_matched.copy()
-    all_pep_cand_list = ref_pep_cand_list.copy()
-    if include_decoys and len(decoy_pep_cand) > 0:
-        all_lib_peaks_matched.extend(decoy_lib_peaks_matched)
-        if 'decoy_pep_cand_list' in locals():
-            all_pep_cand_list.extend(decoy_pep_cand_list)
-        # For decoys, we need to get fragment info from the original (non-decoy) peptides
-        for pep_id in decoy_pep_cand:
-            # Remove "Decoy_" prefix to get original peptide ID
-            if isinstance(pep_id, str) and pep_id.startswith("Decoy_"):
-                orig_id = pep_id[6:]
-            elif isinstance(pep_id, tuple) and pep_id[0].startswith("Decoy_"):
-                orig_id = (pep_id[0][6:], *pep_id[1:])
-            else:
-                orig_id = pep_id
-                
-            if orig_id in decoy_library:
-                if "frags" in decoy_library[orig_id]:
-                    prec_frags.append(decoy_library[orig_id]["frags"])
-                else:
-                    prec_frags.append({})
-                    
-                if "ordered_frags" in decoy_library[orig_id]:
-                    ordered_frags.append(decoy_library[orig_id]["ordered_frags"])
-                else:
-                    ordered_frags.append([])
-            else:
-                prec_frags.append({})
-                ordered_frags.append([])
-    
-    # Pass the original spectrum_matrix (with original indices) and the full processed_spectrum 
-    # to feature calculations, not the ranked/filtered versions
-    
-    # TEMPORARY: Quick test to check if feature calculation is the bottleneck
-    SKIP_HEAVY_FEATURES = True  # Set to False to re-enable full features
-    
-    with Timer("fitting_core.calculate_features"):
-        if SKIP_HEAVY_FEATURES:
-            # Create minimal features for performance testing
-            from .types import SpectralFeatures, BasicFeatures, SimilarityMetrics, StatisticalFeatures, FragmentInfo
-            features = SpectralFeatures(
-                basic=BasicFeatures(
-                    num_peaks_matched=np.ones(len(spectrum_matrix.peptide_candidates)),
-                    frac_lib_intensity=np.ones(len(spectrum_matrix.peptide_candidates)) * 0.5,
-                    total_intensity=np.ones(len(spectrum_matrix.peptide_candidates)),
-                    normalized_intensity=np.ones(len(spectrum_matrix.peptide_candidates)) * 0.5
-                ),
-                similarity=SimilarityMetrics(
-                    cosine_similarity=np.ones(len(spectrum_matrix.peptide_candidates)) * 0.8,
-                    manhattan_distance=np.ones(len(spectrum_matrix.peptide_candidates)) * 0.2,
-                    scribe_score=np.ones(len(spectrum_matrix.peptide_candidates)) * 0.8
-                ),
-                statistical=StatisticalFeatures(
-                    residuals=np.ones(len(spectrum_matrix.peptide_candidates)) * 0.1,
-                    goodness_of_fit=np.ones(len(spectrum_matrix.peptide_candidates)) * 0.9
-                ),
-                fragment_info=FragmentInfo(
-                    frag_errors=[],
-                    frag_names=[],
-                    frag_mz=[],
-                    frag_int=[],
-                    obs_int=[],
-                    unique_frag_mz=[],
-                    unique_obs_int=[]
-                )
-            )
-        else:
-            features = calculate_all_features(
-                spectrum_matrix,  # This has the original row indices
-                processed_spectrum,  # This is the full spectrum
-                lib_coefficients,
-                sparse_matrix,
-                rt_mz,
-                window_idxs,
-                prec_info.rt,
-                np.concatenate([ref_ms1_error, decoy_ms1_error]) if len(decoy_pep_cand) > 0 else ref_ms1_error,
-                ref_spec_offset,
-                decoy_spec_offset,
-                all_lib_peaks_matched,
-                prec_frags,
-                ordered_frags,
-                all_pep_cand_list
-            )
-    
-    return SpectralFitResult(
-        features=features,
-        coefficients=lib_coefficients,
-        peptide_ids=spectrum_matrix.peptide_candidates,
-        is_decoy=spectrum_matrix.is_decoy,
-        sparse_matrix=sparse_matrix,
-        matched_peak_indices=unique_row_idxs
-    )
